Service implementation and interface name change
