+++ src/main/java/com/example/todolist/service/TodoService.java
@@ -1,49 +1,14 @@
 package com.example.todolist.service;
 
 import com.example.todolist.model.Task;
-import com.example.todolist.repository.TodoRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 
-@Service
-@Transactional
-public class TodoService implements TodoServiceInterface {
+public interface TodoService {
 
-    private final TodoRepository todoRepository;
-
-    public TodoService(TodoRepository todoRepository) {
-        this.todoRepository = todoRepository;
-    }
-
-    @Override
-    public List<Task> findAll() {
-        return todoRepository.findAll();
-    }
-
-    @Override
-    public Task findOne(Long id) {
-        return todoRepository.findById(id).orElseThrow(NoSuchElementException::new);
-    }
-
-    @Override
-    public Task save(Task task) {
-        return todoRepository.save(task);
-    }
-
-    @Override
-    public Task update(Task task) {
-        if(todoRepository.findById(task.getId()).isPresent()) {
-            return todoRepository.save(task);
-        } else {
-            throw new NoSuchElementException();
-        }
-    }
-
-    @Override
-    public void delete(Long id) {
-        todoRepository.deleteById(id);
-    }
+    public List<Task> findAll();
+    public Task findOne(Long id);
+    public Task save(Task task);
+    public Task update(Task task);
+    public void delete(Long id);
 }
